pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usb-helper"
dynamic = ["version"]
description = "A Python USB device monitoring and communication framework built on pyusb/libusb"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Yuan-Yi Chang", email = "changyy.csie@gmail.com" },
]
keywords = ["usb", "libusb", "pyusb", "device", "monitor", "hotplug", "bulk", "scsi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Hardware :: Universal Serial Bus (USB)",
]
dependencies = [
    "pyusb>=1.2.1",
    "tomli>=1.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
usb-helper = "usb_helper.cli:main"

[project.urls]
Homepage = "https://github.com/changyy/py-usb-helper"
Repository = "https://github.com/changyy/py-usb-helper"
Issues = "https://github.com/changyy/py-usb-helper/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "usb_helper.__version__"}
